Break up into more modules
* Separate internals from the public API and from each other.

* Add pattern synonyms for understanding `SeqT` as its view.
