+++ include/logict-sequence.h
@@ -1,0 +1,12 @@
+/*
+ * Common macros for logict-sequence
+ */
+
+#ifndef HASKELL_LOGICT_SEQUENCE_H
+#define HASKELL_LOGICT_SEQUENCE_H
+
+#if __GLASGOW_HASKELL__ >= 804
+#define USE_PATTERN_SYNONYMS 1
+#endif
+
+#endif
